feat: add OG metadata for community/blog and canonical; update site url; docs: env var separation guidance
