+++ src/app/[slug]/classes/page.tsx
@@ -34,9 +34,17 @@ export default function CommunityClassesPage() {
 
   if (loading) {
     return (
-      <div className="flex items-center justify-center min-h-screen">
-        <div className="text-slate-500">로딩 중...</div>
-      </div>
+      <main className="pt-14 md:pt-6 pb-3 px-[2%]">
+        <div className="py-12 flex items-center justify-center">
+          <div className="flex items-center gap-2 rounded-2xl bg-black/5 border border-black/10 px-3 py-2 shadow-sm">
+            <svg className="w-5 h-5 animate-spin text-slate-800" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+              <circle className="opacity-20" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
+              <path d="M4 12a8 8 0 018-8" stroke="currentColor" strokeWidth="4" strokeLinecap="round" />
+            </svg>
+            <span className="text-sm font-medium text-slate-800">불러오는 중...</span>
+          </div>
+        </div>
+      </main>
     )
   }
